backend/requirements.txt
alembic==1.4.0
amqp==2.5.2
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
billiard==3.6.2.0
celery==4.3.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.7
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
redis==3.3.11
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
vine==1.3.0
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
zipp==2.2.0
